+++ CHANGELOG.md
@@ -3,13 +3,15 @@
 All notable changes to this project are documented in this file.
 
 ## [Unreleased]
+
+## [0.2.2] - 2018-06-21
 ### Added
-- ProgramTarget in genairics.targets, for program dependencies.
+- ProgramTarget in genairics.targets and ProgramDependencyBase in
+##genairics.tasks, for program dependencies.
 - --cluster-PPN option for qsub job-launcher to specify number of
 processors per node to request.
 - Created genairics.targets, genairics.tasks and genairics.config.
 ### Changed
-- runWorkflow now only prints out task info if --verbose
 - coverageTask to RNAseq pipeline to generate bw coverage
 file. this also sorts the bam and creates its index for
 visualization in a genome browser.